requirements.txt
pyzmq==17.0.0
fire==0.1.3
redis
# PyInstaller==3.3.1
PyYAML==3.12
# guizero==0.4.5
